test: add unit tests for configuration loading and server setup, improving error handling and coverage
